refactor: rename functions
